move utils and web3 state
